+++ monocle/tests/unit/test_api.py
@@ -18,7 +18,7 @@ import logging
 import unittest
 from flask import json
 from monocle import webapp
-from monocle import projects
+from monocle import config
 from monocle.db.db import ELmonocleDB
 
 from .common import index_dataset
@@ -53,12 +53,12 @@ class TestWebAPI(unittest.TestCase):
         webapp.CHANGE_PREFIX = self.prefix
         webapp.app.config['TESTING'] = True
         self.client = webapp.app.test_client()
-        projects_data = {
-            "projects": [
+        config_data = {
+            "tenants": [
                 {
                     # Private index
                     "index": self.index1,
-                    "users_whitelist": ['jane', 'john'],
+                    "users": ['jane', 'john'],
                 },
                 {
                     # Public index
@@ -66,7 +66,7 @@ class TestWebAPI(unittest.TestCase):
                 },
             ]
         }
-        webapp.indexes_acl = projects.build_index_acl(projects_data)
+        webapp.indexes_acl = config.build_index_acl(config_data)
 
     def test_get_indices(self):
         "Test indices endpoint"